requirements.txt
django>=2.1,<=2.2
djangorestframework
markdown
django-filter
